#96 move new meta model object to new package
