+++ api/parser/src/main/scala/de/htwg/zeta/parser/GraphicalDSLParser.scala
@@ -4,9 +4,9 @@ import scalaz.Failure
 import scalaz.Success
 import scalaz.Validation
 
-import de.htwg.zeta.common.model.MetaModel
 import de.htwg.zeta.common.model.shape.Shape
 import de.htwg.zeta.common.models.project.concept.Concept
+import de.htwg.zeta.common.models.project.gdsl.MetaModel
 import de.htwg.zeta.parser.diagram.DiagramParseTreeTransformer
 import de.htwg.zeta.parser.diagram.DiagramParser
 import de.htwg.zeta.parser.shape.ShapeParseTreeTransformer
